requirements.txt
Pillow
pycairo
opencv-python==4.5.5.62
